+++ roomba.lua
@@ -1,6 +1,6 @@
--- Roomba Hive command utility v0.2.1
+-- Roomba Hive command utility v0.2.2
 
-local VERSION = "0.2.1"
+local VERSION = "0.2.2"
 local args = { ... }
 local command = args[1] and args[1]:lower() or "help"